requirements.txt
tensorflow==2.10.1
keras==2.10.0
pillow
fastapi
uvicorn